refactor(node/sync): extract handshake submodule from the manager god-file (increment 3)
Continues the sync/manager split. Moves the sync-handshake construction
cluster into `handshake.rs` — behavior-preserving, pure code movement
(mod.rs: block removed + module wiring; also drops the now-unused
build_handshake_from_raw / estimate_entity_count / estimate_max_depth /
SyncHandshake imports the moved methods owned).

Moved (an `impl SyncManager` fragment):
- build_local_handshake, build_remote_handshake (pub(super); called from the
  protocol-negotiation flow in mod.rs)
- query_tree_stats (private; used only within the cluster)

Explicit imports, no `use super::*`. Methods staying in mod.rs remain
reachable via ancestor privacy.

Validated: node lib tests (333) and sync_sim (314) pass.
